+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## [0.1.0] - 2023-05-23
 
 - Initial release
+  - Added `GooglePalmApi::Client` class that supports `embed()`, `generate_text()`, `generate_chat_message()`, `get_model()`, `list_models()` and `count_message_tokens()` methods